Moved claimant tests into subfolder
