+++ src/applications/representative-appoint/tests/pages/claimant/claimantContactMailing.unit.spec.jsx
@@ -5,7 +5,7 @@ import { render } from '@testing-library/react';
 import { DefinitionTester } from 'platform/testing/unit/schemaform-utils';
 import { $ } from 'platform/forms-system/src/js/utilities/ui';
 
-import formConfig from '../../config/form';
+import formConfig from '../../../config/form';
 
 describe('Claimant Contact Mailing page', () => {
   const {
